backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
